feat: adicionado pasta do strategy e padrao implementado
