+++ backend/americanas/models/__init__.py
@@ -5,5 +5,3 @@ from .EnderecoUsuario import EnderecoUsuario
 from .Produto import Produto
 from .Avaliacao import Avaliacao
 from .Contem import ContemProdutoPedido
-from .Avaliacao import CalculadoraMetrica
-from .Avaliacao import MediaPonderadaStrategy
