+++ backend/strategy/__init__.py
@@ -1,0 +1,1 @@
+from .base_strategy import CalculadoraMetrica, MediaPonderadaStrategy
