+++ backend/strategy/base_strategy.py
@@ -1,0 +1,20 @@
+from abc import ABC, abstractmethod
+# Interface para a estratégia de cálculo de métricas
+class CalculoMetricaStrategy(ABC):
+    @abstractmethod
+    def calcular_metrica(self, avaliacao):
+        pass
+
+# Estratégia concreta para calcular a média ponderada
+class MediaPonderadaStrategy(CalculoMetricaStrategy):
+    def calcular_metrica(self, avaliacao):
+        peso = 1  # Defina o peso conforme necessário
+        return (avaliacao.nota * peso) / peso
+
+# Classe que utiliza a estratégia para calcular a métrica
+class CalculadoraMetrica:
+    def __init__(self, estrategia):
+        self.estrategia = estrategia
+
+    def calcular(self, avaliacao):
+        return self.estrategia.calcular_metrica(avaliacao)
